Updated Health.cs
Separated health functions into its own script, updated meleeenemy prefab, updated creature script
