+++ Assets/Scripts/Health.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73327dba548c99946b8676136f0a3463
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
