+++ Assets/Scripts/MeleeEnemy.cs
@@ -14,14 +14,7 @@ public class MeleeEnemy : MonoBehaviour
 
     private float cooldownTimer = Mathf.Infinity;
 
-    //temp
     public Health playerHealth;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
 
     // Update is called once per frame
     void Update()
@@ -40,15 +33,22 @@ public class MeleeEnemy : MonoBehaviour
 
     public bool PlayerInSight()
     {
-        RaycastHit2D hit = Physics2D.BoxCast(boxCollider.bounds.center + transform.right * range * transform.localScale.x * colliderDistance, 
-        new Vector3(boxCollider.bounds.size.x * range, boxCollider.bounds.size.y, boxCollider.bounds.size.z)
-        , 0, Vector2.left, 0, playerLayer);
+        RaycastHit2D hit = PerformBoxCast();
 
         if (hit.collider != null)
         {
             playerHealth = hit.transform.GetComponent<Health>();
         }
+
         return hit.collider != null;
+    }
+
+    private RaycastHit2D PerformBoxCast()
+    {
+        Vector2 castOrigin = boxCollider.bounds.center + transform.right * range * transform.localScale.x * colliderDistance;
+        Vector3 castSize = new Vector3(boxCollider.bounds.size.x * range, boxCollider.bounds.size.y, boxCollider.bounds.size.z);
+
+        return Physics2D.BoxCast(castOrigin, castSize, 0, Vector2.left, 0, playerLayer);
     }
 
     private void OnDrawGizmos()
