+++ Assets/Scripts/creature.cs
@@ -6,11 +6,10 @@ public class creature : MonoBehaviour
 {
     [Header("Config")]
     public int health;
-    public int maxHealth = 3;  //TODO: change to 100 and update HealthDisplay .. % 10
+    public int maxHealth = 3;  
     public float speed = 6.0f;
     public float jumpForce = 35f;
     public string creatureName = "Lonk";
-    private float DMG_DELAY = 1.5f;
 
     [Header("Projectiles")]
     public GameObject projectile;
@@ -30,7 +29,6 @@ public class creature : MonoBehaviour
     void Start()
     {
         Debug.Log("start called");
-        health = maxHealth;
     }
 
     // Update is called once per frame
@@ -66,53 +64,4 @@ public class creature : MonoBehaviour
         //newProjectile.GetComponent<projectile>().LaunchProjectile(Camera.main.ScreenToWorldPoint(Input.mousePosition));
     }
 
-    /*Health Functions*/
-    public void TakeDamage(int d) 
-    {
-        Debug.Log("took damage");
-        if (health <= 0)
-        {
-            Debug.Log("You died");
-            //TODO: You Died text overlay
-        } else {
-            health -= d;
-        }
-    }
-
-    public void GainHealth(int h)
-    {
-        //play sound effect
-        health += h;
-    }
-
-    /*Collisions / Triggers*/
-    private void OnCollisionEnter2D(Collision2D collision)
-    {
-        if (collision.gameObject.tag == "Enemy")
-        {
-            TakeDamage(1); //temp
-            StartCoroutine(DamageDelay());
-        }
-    }
-
-    private void OnTriggerEnter2D(Collider2D collider)
-    {
-        GameObject collidedObject = collider.gameObject;
-
-        if (collidedObject.tag != "Item")
-            return;
-
-        //switch statement if there are more items
-
-        if (health != maxHealth)
-        {
-            GainHealth(1); //temp value
-            Destroy(collider.gameObject);
-        }
-    }
-
-    IEnumerator DamageDelay()
-    {
-        yield return new WaitForSeconds(DMG_DELAY);
-    }
 }
